make l1-to-l2 test parallel
