--- yarn-project/end-to-end/src/e2e_cross_chain_messaging/l1_to_l2.test.ts
@@ -1,304 +1,0 @@
-import { AztecAddress } from '@aztec/aztec.js/addresses';
-import { generateClaimSecret } from '@aztec/aztec.js/ethereum';
-import { Fr } from '@aztec/aztec.js/fields';
-import type { Logger } from '@aztec/aztec.js/log';
-import { isL1ToL2MessageReady } from '@aztec/aztec.js/messaging';
-import type { AztecNode } from '@aztec/aztec.js/node';
-import { TxExecutionResult } from '@aztec/aztec.js/tx';
-import type { Wallet } from '@aztec/aztec.js/wallet';
-import { BlockNumber, IndexWithinCheckpoint } from '@aztec/foundation/branded-types';
-import { timesAsync } from '@aztec/foundation/collection';
-import { retryUntil } from '@aztec/foundation/retry';
-import { TestContract } from '@aztec/noir-test-contracts.js/Test';
-import { ExecutionPayload } from '@aztec/stdlib/tx';
-
-import { PIPELINING_SETUP_OPTS } from '../fixtures/fixtures.js';
-import { sendL1ToL2Message } from '../fixtures/l1_to_l2_messaging.js';
-import type { CrossChainTestHarness } from '../shared/cross_chain_test_harness.js';
-import { CrossChainMessagingTest } from './cross_chain_messaging_test.js';
-
-describe('e2e_cross_chain_messaging l1_to_l2', () => {
-  let t: CrossChainMessagingTest;
-  let log: Logger;
-  let crossChainTestHarness: CrossChainTestHarness;
-  let aztecNode: AztecNode;
-  let wallet: Wallet;
-  let user1Address: AztecAddress;
-  let testContract: TestContract;
-
-  beforeEach(async () => {
-    t = new CrossChainMessagingTest(
-      'l1_to_l2',
-      // PIPELINING_SETUP_OPTS sets minTxsPerBlock=0; this test needs minTxsPerBlock=1 because it
-      // manually mines blocks one tx at a time via advanceBlock() and counts checkpoints, so empty
-      // pipelined checkpoints interleaving between txs would break the exact block-number assertions.
-      { ...PIPELINING_SETUP_OPTS, minTxsPerBlock: 1 },
-      { aztecProofSubmissionEpochs: 2, aztecEpochDuration: 4 },
-    );
-    await t.setup();
-
-    ({ logger: log, crossChainTestHarness, wallet, user1Address, aztecNode } = t);
-    ({ contract: testContract } = await TestContract.deploy(wallet).send({ from: user1Address }));
-  }, 300_000);
-
-  afterEach(async () => {
-    await t.teardown();
-  });
-
-  const getConsumeMethod = (scope: 'private' | 'public') =>
-    scope === 'private'
-      ? testContract.methods.consume_message_from_arbitrary_sender_private
-      : testContract.methods.consume_message_from_arbitrary_sender_public;
-
-  // Sends a tx to L2 to advance the block number by 1
-  const advanceBlock = async () => {
-    const block = await aztecNode.getBlockNumber();
-    log.warn(`Sending noop tx at block ${block}`);
-    await wallet.sendTx(ExecutionPayload.empty(), { from: user1Address });
-    const newBlock = await aztecNode.getBlockNumber();
-    log.warn(`Advanced to block ${newBlock}`);
-    if (newBlock === block) {
-      throw new Error(`Failed to advance block ${block}`);
-    }
-    // Under interval mining `AnvilTestWatcher.markAsProven` does not auto-fire; without an explicit
-    // prove call here, L1's `aztecProofSubmissionEpochs=2` window (96s with pipelined 12s slots)
-    // expires mid-test and triggers a chain prune that drops in-flight wallet txs.
-    await t.context.watcher.markAsProven();
-    return newBlock;
-  };
-
-  const waitForBlockToCheckpoint = async (blockNumber: BlockNumber) => {
-    return await retryUntil(
-      async () => {
-        const checkpointedBlockNumber = await aztecNode.getBlockNumber('checkpointed');
-        const isCheckpointed = checkpointedBlockNumber >= blockNumber;
-        if (!isCheckpointed) {
-          return undefined;
-        }
-        const [checkpointedBlock] = await aztecNode.getBlocks(blockNumber, 1, {
-          includeL1PublishInfo: true,
-          includeAttestations: true,
-          onlyCheckpointed: true,
-        });
-        return checkpointedBlock.checkpointNumber;
-      },
-      'wait for block to checkpoint',
-      60,
-    );
-  };
-
-  const advanceCheckpoint = async () => {
-    let checkpoint = await aztecNode.getCheckpointNumber();
-    const originalCheckpoint = checkpoint;
-    log.warn(`Original checkpoint ${originalCheckpoint}`);
-    do {
-      const newBlock = await advanceBlock();
-      checkpoint = await waitForBlockToCheckpoint(newBlock);
-    } while (checkpoint <= originalCheckpoint);
-    log.warn(`At checkpoint ${checkpoint}`);
-  };
-
-  // Same as above but ignores errors. Useful if we expect a prune.
-  const tryAdvanceBlock = async () => {
-    try {
-      await advanceBlock();
-    } catch (err) {
-      log.warn(`Failed to advance block: ${(err as Error).message}`);
-    }
-  };
-
-  // Waits until the message is fetched by the archiver of the node and returns the msg target checkpoint
-  const waitForMessageFetched = async (msgHash: Fr) => {
-    log.warn(`Waiting until the message is fetched by the node`);
-    return await retryUntil(
-      async () => {
-        const checkpoint = await aztecNode.getL1ToL2MessageCheckpoint(msgHash);
-        if (checkpoint !== undefined) {
-          return checkpoint;
-        }
-        await advanceBlock();
-        return undefined;
-      },
-      'get msg checkpoint',
-      60,
-    );
-  };
-
-  // Waits until the message is ready to be consumed on L2 as it's been added to the world state
-  const waitForMessageReady = async (
-    msgHash: Fr,
-    scope: 'private' | 'public',
-    onNotReady?: (blockNumber: BlockNumber) => Promise<void>,
-  ) => {
-    const msgCheckpoint = await waitForMessageFetched(msgHash);
-    log.warn(
-      `Waiting until L2 reaches the first block of msg checkpoint ${msgCheckpoint} (current is ${await aztecNode.getCheckpointNumber()})`,
-    );
-    await retryUntil(
-      async () => {
-        const [blockNumber, checkpointNumber] = await Promise.all([
-          aztecNode.getBlockNumber(),
-          aztecNode.getCheckpointNumber(),
-        ]);
-        const witness = await aztecNode.getL1ToL2MessageMembershipWitness('latest', msgHash);
-        const isReady = await isL1ToL2MessageReady(aztecNode, msgHash);
-        log.info(
-          `Block is ${blockNumber}, checkpoint is ${checkpointNumber}. Message checkpoint is ${msgCheckpoint}. Witness ${!!witness}. Ready ${isReady}.`,
-        );
-        if (!isReady) {
-          await (onNotReady ? onNotReady(blockNumber) : advanceBlock());
-        }
-        return isReady;
-      },
-      `wait for rollup to reach msg checkpoint ${msgCheckpoint}`,
-      240,
-    );
-  };
-
-  // We register one portal address when deploying contract but that address is no-longer the only address
-  // allowed to send messages to the given contract. In the following test we'll test that it's really the case.
-  // We'll also test that we can send the same message content across the bridge multiple times.
-  it.each(['private', 'public'] as const)(
-    'can send an L1 to L2 message from a non-registered portal address consumed from %s repeatedly',
-    async (scope: 'private' | 'public') => {
-      // Generate and send the message to the L1 contract
-      const [secret, secretHash] = await generateClaimSecret();
-      const message = { recipient: testContract.address, content: Fr.random(), secretHash };
-      const { msgHash: message1Hash, globalLeafIndex: actualMessage1Index } = await sendL1ToL2Message(
-        message,
-        crossChainTestHarness,
-      );
-
-      await waitForMessageReady(message1Hash, scope);
-
-      const [message1Index] = (await aztecNode.getL1ToL2MessageMembershipWitness('latest', message1Hash))!;
-      expect(actualMessage1Index.toBigInt()).toBe(message1Index);
-
-      const sendConsumeMsgTx = async (index: Fr) => {
-        const call = getConsumeMethod(scope)(message.content, secret, crossChainTestHarness.ethAccount, index);
-        if (scope === 'public') {
-          await call.simulate({ from: user1Address });
-        }
-        await call.send({ from: user1Address });
-      };
-
-      // We consume the L1 to L2 message using the test contract either from private or public
-      await sendConsumeMsgTx(actualMessage1Index);
-
-      // We send and consume the exact same message the second time to test that oracles correctly return the new
-      // non-nullified message
-      const { msgHash: message2Hash, globalLeafIndex: actualMessage2Index } = await sendL1ToL2Message(
-        message,
-        crossChainTestHarness,
-      );
-
-      // We check that the duplicate message was correctly inserted by checking that its message index is defined
-      await waitForMessageReady(message2Hash, scope);
-
-      const [message2Index] = (await aztecNode.getL1ToL2MessageMembershipWitness('latest', message2Hash))!;
-      expect(message2Index).toBeDefined();
-      expect(message2Index).toBeGreaterThan(actualMessage1Index.toBigInt());
-      expect(actualMessage2Index.toBigInt()).toBe(message2Index);
-
-      // Now we consume the message again. Everything should pass because oracle should return the duplicate message
-      // which is not nullified
-      await sendConsumeMsgTx(actualMessage2Index);
-    },
-    300_000,
-  );
-
-  // Inbox checkpoint number can drift on two scenarios: if the rollup reorgs and rolls back its own
-  // checkpoint number, or if the inbox receives too many messages and they are inserted faster than
-  // they are consumed. In this test, we mine several checkpoints without marking them as proven until
-  // we can trigger a reorg, and then wait until the message can be processed to consume it.
-  it.each(['private', 'public'] as const)(
-    'can consume L1 to L2 message in %s after inbox drifts away from the rollup',
-    async (scope: 'private' | 'public') => {
-      // Reset the L1 proof window by marking the current pending tip as proven. The e2e fixture
-      // runs L1 on interval mining, so the watcher's auto-prove loop never starts (it gates on
-      // `isAutoMining`). That means L1's prune deadline has been anchored to chain genesis the
-      // whole setup, and would otherwise fire mid-test before we finish mining the 4 drift
-      // checkpoints below.
-      await t.context.watcher.markAsProven();
-
-      // Stop proving
-      const lastProven = await aztecNode.getBlockNumber();
-      const [checkpointedProvenBlock] = await aztecNode.getBlocks(lastProven, 1, {
-        includeL1PublishInfo: true,
-        includeAttestations: true,
-        onlyCheckpointed: true,
-      });
-      log.warn(`Stopping proof submission at checkpoint ${checkpointedProvenBlock.checkpointNumber} to allow drift`);
-      t.context.watcher.setIsMarkingAsProven(false);
-
-      // Mine several checkpoints to ensure drift
-      log.warn(`Mining blocks to allow drift`);
-      await timesAsync(4, advanceCheckpoint);
-
-      // Generate and send the message to the L1 contract
-      log.warn(`Sending L1 to L2 message`);
-      const [secret, secretHash] = await generateClaimSecret();
-      const message = { recipient: testContract.address, content: Fr.random(), secretHash };
-      const { msgHash, globalLeafIndex } = await sendL1ToL2Message(message, crossChainTestHarness);
-
-      // Wait until the Aztec node has synced it
-      const msgCheckpointNumber = await waitForMessageFetched(msgHash);
-      log.warn(`Message synced for checkpoint ${msgCheckpointNumber}`);
-      expect(checkpointedProvenBlock.checkpointNumber + 4).toBeLessThan(msgCheckpointNumber);
-
-      // And keep mining until we prune back to the original block number. Now the "waiting for two blocks"
-      // strategy for the message to be ready to use shouldn't work, since the lastProven block is more than
-      // two blocks behind the message block. This is the scenario we want to test.
-      log.warn(`Waiting until we prune back to ${lastProven}`);
-      await retryUntil(
-        async () =>
-          (await aztecNode.getBlockNumber().then(b => b === lastProven || b === lastProven + 1)) ||
-          (await tryAdvanceBlock()),
-        'wait for prune',
-        180,
-      );
-
-      // Check that there is no witness yet
-      expect(await aztecNode.getL1ToL2MessageMembershipWitness('latest', msgHash)).toBeUndefined();
-
-      // Define L2 function to consume the message
-      const consume = () =>
-        getConsumeMethod(scope)(message.content, secret, crossChainTestHarness.ethAccount, globalLeafIndex);
-
-      // Wait until the message is ready to be consumed, checking that it cannot be consumed beforehand
-      await waitForMessageReady(msgHash, scope, async () => {
-        if (scope === 'private') {
-          // On private, we simulate the tx locally and check that we get a missing message error, then we advance to the next block
-          await expect(() => consume().simulate({ from: user1Address })).rejects.toThrow(/No L1 to L2 message found/);
-          await tryAdvanceBlock();
-        } else {
-          // In public it is harder to determine when a message becomes consumable.
-          // We send a transaction, this advances the chain and the message MIGHT be consumed in the new block.
-          // If it does get consumed then we check that the block contains the message.
-          // If it fails we check that the block doesn't contain the message
-          const { receipt } = await consume().send({ from: user1Address, wait: { dontThrowOnRevert: true } });
-          if (receipt.executionResult === TxExecutionResult.SUCCESS) {
-            // The block the transaction included should be for the message checkpoint number
-            // and be the first block in the checkpoint
-            const block = await aztecNode.getBlock(receipt.blockNumber!);
-            expect(block).toBeDefined();
-            expect(block!.checkpointNumber).toEqual(msgCheckpointNumber);
-            expect(block!.indexWithinCheckpoint).toEqual(IndexWithinCheckpoint.ZERO);
-          } else {
-            expect(receipt.executionResult).toEqual(TxExecutionResult.REVERTED);
-          }
-        }
-        await t.context.watcher.markAsProven();
-      });
-
-      // Verify the membership witness is available for creating the tx (private-land only)
-      if (scope === 'private') {
-        const [messageIndex] = (await aztecNode.getL1ToL2MessageMembershipWitness('latest', msgHash))!;
-        expect(messageIndex).toEqual(globalLeafIndex.toBigInt());
-        // And consume the message for private, public was already consumed.
-        await consume().send({ from: user1Address });
-      }
-    },
-    300_000,
-  );
-});
